Move ColorManager to ui
Part of #2005
Part of #1507
