+++ src/renderer/TextRenderLayer.ts
@@ -3,13 +3,14 @@
  * @license MIT
  */
 
-import { IColorSet, IRenderDimensions, ICharacterJoinerRegistry } from './Types';
+import { IRenderDimensions, ICharacterJoinerRegistry } from './Types';
 import { ITerminal } from '../Types';
 import { CharData, ICellData } from '../core/Types';
 import { GridCache } from './GridCache';
 import { BaseRenderLayer } from './BaseRenderLayer';
 import { CellData, AttributeData, Content, NULL_CELL_CODE } from '../core/buffer/BufferLine';
 import { JoinedCellData } from './CharacterJoinerRegistry';
+import { IColorSet } from '../ui/Types';
 
 /**
  * This CharData looks like a null character, which will forc a clear and render
