+++ src/renderer/atlas/CharAtlasCache.ts
@@ -4,13 +4,13 @@
  */
 
 import { ITerminal } from '../../Types';
-import { IColorSet } from '../Types';
 import { generateConfig, configEquals } from './CharAtlasUtils';
 import BaseCharAtlas from './BaseCharAtlas';
 import DynamicCharAtlas from './DynamicCharAtlas';
 import NoneCharAtlas from './NoneCharAtlas';
 import StaticCharAtlas from './StaticCharAtlas';
 import { ICharAtlasConfig } from './Types';
+import { IColorSet } from '../../ui/Types';
 
 const charAtlasImplementations = {
   'none': NoneCharAtlas,
