+++ src/renderer/atlas/CharAtlasGenerator.ts
@@ -5,8 +5,8 @@
 
 import { FontWeight } from 'xterm';
 import { isFirefox, isSafari } from '../../common/Platform';
-import { IColor } from '../Types';
 import { ICharAtlasConfig, CHAR_ATLAS_CELL_SPACING } from './Types';
+import { IColor } from '../../ui/Types';
 
 /**
  * Generates a char atlas.
