+++ src/renderer/atlas/CharAtlasUtils.ts
@@ -4,9 +4,9 @@
  */
 
 import { ITerminal } from '../../Types';
-import { IColorSet } from '../Types';
 import { ICharAtlasConfig } from './Types';
 import { DEFAULT_COLOR } from '../../common/Types';
+import { IColorSet } from '../../ui/Types';
 
 export function generateConfig(scaledCharWidth: number, scaledCharHeight: number, terminal: ITerminal, colors: IColorSet): ICharAtlasConfig {
   // null out some fields that don't matter
