+++ src/ui/ColorManager.test.ts
